fix(KaotoToolbar): Focus input when editing integration name. #768
Currently, when clicking on the edit button labeled with a pencil icon,
enables the edit mode but doesn't focus on the input field.

This commit adds a callback to be executed whenever the input field appears
to set focus on it and avoid the need to click on it to start editing.