fix(MiniCatalog): Set a fixed id to Popover to avoid recalculate it
At the moment, every time that a Popover is rendered, an id is
calculated if not provided.

This lead to the situation that if something in the DOM tree is updated,
the Popover will get re-rendered but now [with a new id](https://github.com/patternfly/patternfly-react/blob/d526d77ef63c407a7917657536c4d242822e956b/packages/react-core/src/components/Popover/Popover.tsx#L272),
which in turn causes a blocking refresh in the VSCode extension.

This commit adds a fixed id to the VisualizationStep component to
remediate that behavior.

Relates https://github.com/KaotoIO/kaoto-ui/issues/1316