chore: Remove store dependencies from VisualizationService and StepsService
At the moment, in order to work with the VisualizationService and
StepsService, binding stores to each component is required.

The downside of this is that in case of any property chance, the
component will rerender.

The goal for this commit is to first, deprecate the contructor's
parameters from the aforementioned services and secondly, use the
getState() utility method in order to retrieve the latest without
binding the store.

Relates to https://github.com/KaotoIO/kaoto-ui/issues/1314