fix(viz): Avoid rerender VisualizationStepViews component
At the moment, whenever the step's detail view is open,
hovering over a step forces the VisualizationStepViews to
rerender creating a content flash.

Built on top of the previous commit, this commit removes the
need of binding entire stores in the following components:

- Visualization
- VisualizationStepViews

and only subscribes to the individual properties that needs to
be watched.

Fixes [1314](https://github.com/KaotoIO/kaoto-ui/issues/1314)