fix(Visualization): Sync Code editor with step's details view
Selecting a given step, it populates a selectedStep variable which it's used
in combination with the isPanelExpanded variable to display the step's details.

The problem comes when a change to said step happens outside of the details view
because those changes are not propagated to the VisualizationStepViews component.

A fix for this problem is to listen to selectedStep.UUID property in combination
with the integrationJson property to reload the information whenever
a change happens.

Fixes https://github.com/KaotoIO/kaoto-ui/issues/1395