fix(viz): Avoid displaying the MiniCatalog if there are no steps or branches available
At this moment, trying to append an additional step after a step that has a
custom extension, causes to display the MiniCatalog with both tabs disabled
and showing the Add branch button enabled.

This is problematic because for those steps, a special configuration is
required and its provided via its extension, hence disabling the branch
tab from the MiniCatalog.

This commit disables the plus icon to avoid displaying a non interactive
popup to the user and providing a tooltip instead explaining what to do
in order to configure the step.

This commit also target small improvements in related tests, like changing
the api signature of one of the services to only receive a step instead of
a whole node.

fixes https://github.com/KaotoIO/kaoto-ui/issues/1473