chore: Replace Prepend/Append component's template conditional with CSS
In case that the layout from ReactFlow change, we need to update the
handlers and Prepend/Append buttons as well. Currently this is being
handled with a template conditional to determine which CSS class
should be applied.

This commit aims to simplify that by leveraging the concept of
specificity by relying on a class applied to the parent element
and the children reacting to that, this way we rely on the browser
to do the changes instead of JS.

This is a spinoff of https://github.com/KaotoIO/kaoto-ui/pull/1485/files#r1138269126