fix(step-details): Sort step's parameters
At the moment, we're using the same order provided from the backend
to present the step's parameter list to the user. Sometimes this is
problematic because for some steps, the required fields are at the
end of the list.

This commit sort the step's parameter list using the following criteria:

    1. Move mandatory parameters to the top
    2. Sort them alphabetically
    3. Sort alphabetically the rest of the parameters

fixes https://github.com/KaotoIO/kaoto-ui/issues/1301