fix(step-details): Debounce saveConfig callback
At the moment of writing, whenever an user writes a character, the
saveConfig callback is executed and 2 HTTP calls are triggered to
update the YAML integration file.

This becomes problematic for bigger integrations or for long strings
because triggers unnecessary server calls.

This commit debounces that call by 500ms to leave time to the user
to finish writing before updating the YAML file.

fixes https://github.com/KaotoIO/kaoto-ui/issues/1339