fix(vscode): Sort Step's parameters in VS Code
Originally, the step's parameters order was based on the premise that
the properties order is kept when going from Array -> Object in modern
browsers, but also that using the ApiService was the only place in which
a step detail was fetched.

Both assumptions were not completely correct, so now the sort strategy
it's based on the AutoForm library which allows specifying a fixed order
for the automated generated fields.

fixes https://github.com/KaotoIO/kaoto-ui/issues/1301 (again)