fix(StepsService): Generate steady IDs for steps and branches
At the moment, when a step is no longer available because it was removed,
or its id changed, we close the step's detail panel since is no longer
editing the right step.

This causes an issue with vscode since there's a circular dependency
between the integration JSON and the code editor. Long story short,
whenever there's an update of the integration JSON, the steps or
branch IDs change, forcing a new sync cycle from the kogito integration side.

A partial fix to this issue is to ensure that the IDs are consistent between syncs,
so this way the comparison inside of the KogitoEditorIntegrationProvider.tsx works.

fix: https://github.com/KaotoIO/kaoto-ui/issues/1553