fix(step-details): Step's details panel is referencing an old value
Currently, when opening the step’s details tab, it’s not referencing the
most recent values but keeping the previous values from a previous step.

This generates unintentional changes to a different step than the user is
intending to.

This situation was caused because of the addition of a `debounce` function
with the intention to only save the
[form’s values after the user’s input stopped](https://github.com/KaotoIO/kaoto-ui/pull/1537).

The problem was that the `debounce` function was causing to capture the form’s
value from the previous render, hence saving the wrong values to the previous step.

To avoid this situation in the future, the saveConfig function now
receives the original step and the new values.

- Changes
* Refactor saveConfig callback to receive the original step as parameter
* Replace lodash' debounce function with useDebounceCallback()
* Increase the debounce timer from 500ms to 1s
* Update onChangeModel signature to (configuration: Record<string, unknown>, isValid: boolean)
  to reflect more accurately the intended parameters.
* Remove the lodash.debounce dependency since is no longer required

fix: https://github.com/KaotoIO/kaoto-ui/issues/1561