chore: Improve zustand mock
If createState is empty, store won't have a getState function
and it will be the real createImpl function.

In this situation, we need to make sure to return the function
itself to be able to intercept the original createState function
and be able to reset the store after each test.