feat(api): Allow configuring the API URL at runtime
Currently, the API URL is retrieved using an environment variable
available at build time only. This creates friction in the VSCode
extension context as it blocks running more than one instance of it.

This commit adds a constructor parameter to optionally receive the
API URL so it can be overridden through the VSCode bootstrap process.

The API URL still defaults to the same environment variable provided
at build time.

fixes: [1451](https://github.com/KaotoIO/kaoto-ui/issues/1451)