chore: rename appendStep to insertStep
In the existing useIntegrationJSONStore, there are individual methods
to handle the different interaction between the flows' steps.

In the new useFlowsStore the attempt is to have a single method that
handle the same interactions but using a command-like approach of
'append', 'insert', 'replace'.

This API might be subject to change in the future as more functionality
gets ported from single flow to multiple flows

Relates to: https://github.com/KaotoIO/kaoto-ui/issues/801