feat(multiple-flows): Enable experimental support for multiple flows
In order to support multiple flows in the UI, we need to switch our
model from a single flow to an array of flows.

In practical terms, there are at least two alternatives:
1. Refactor existing `IIntegrationJSONStore` to support multiple flows
2. Move from `IIntegrationJSONStore` to another store that supports
multiple flows.

In this commit, the latter was chosen, and for that purpose, the `useFlowsStore`
was created.

In this commit, also an additional setting was added to switch between single and
multiple flows during runtime, to avoid any potential big-bang effect while
developing this feature.

Relates to: https://github.com/KaotoIO/kaoto-ui/issues/801