feat: Support editing steps from multiple flows
Currently we're calculating the step's ids by leveraging object references, this come
handy for branches but presents the problem of leaking object references between single
and multiple flows.

This commit leverages lodash.clonedeep dependency to make a copy of the steps as a
workaround for those situations.

This should change in the future with a more robust solution like immer.

relates to: https://github.com/KaotoIO/kaoto-ui/issues/801