fix(stepsService): Use id instead of title for steps parameters
In the past, the 'id' and 'title' properties from a step were the same.
With the introduction of https://github.com/KaotoIO/kaoto-backend/pull/650, now
those properties are not the same.

On the UI side, the 'title' property was used to build the parameters of the step, so
in combination with the previous fix, this caused step extensions relying upon
on the previous information are no longer working.

The fix is to use the 'id' property form now on

relates to: https://github.com/KaotoIO/kaoto-ui/issues/1534