chore: Remove IntegrationJSONStore
After merging the last pull request about enabling
the flowsStore, the next step is to remove the previous
integrationJsonStore.

In addition to that, another needed step is to remove the
temporary settings switch that toggles between the previous
and the new store.

This pull request removes the integrationJsonStore and adjust
the existing unit tests to comply with the new store.

relates to: https://github.com/KaotoIO/kaoto-ui/issues/801