fix(chromatic): Fix validation errors that shows in chromatic
Currently the "chromatic" build step is tagging all commits as failed
since there are a few errors showing up as now all endpoints are being
mocked properly.

This pull request adds a small validation until we have a better mocked
setup to support all components and their stories.

It also removes the SourceCodeEditor story as we're not providing the
necessary http mocks to ensure that works properly.