feat(flows): Add mechanism to navigate between flows
Currently, all existing flows are being shown. This behavior
is not useful when there are many flows at the same time or
somewhat complex flows.

In addition to that, there's no mechanism to delete an existing
flow.

This commit provides a way to toggle the visibility of individual
flows, by leveraging a flows list that also provides the functionality
of deleting individual flows.

relates to: https://github.com/KaotoIO/kaoto-ui/issues/801