fix(viz): Avoid rerendering of VisualizationStepViews
Currently, the Visualization component is consuming several properties
from the VisualizationStore, and this makes it more susceptible to
store changes, causing rerendering of the VisualizationStepViews
component, and this in turn generates a flickering in the step
extensions.

The fix is to add the 'shallow' function to the 'useVisualizationStore()'
in order to only react to changes in the actual value of said properties.

Relates to: https://github.com/KaotoIO/kaoto-ui/issues/801