chore(tests): Adjust react package resolution for @kie-tools/uniforms-patternfly
With the upgrade of yarn to v3, JsonSchemaConfigurator.test is failing since now
yarn have a better version enforcing. This leads @kie-tools/uniforms-patternfly
to have their own react and react-dom version (^17.0.2) and this causes a mistmatch
with the project react and react-dom version (18.2.0).

The workaround for this situation is to add a package version resolution for
react and react-dom to ensure that all dependencies are using the same version.