fix(vscode): Routes gets reset upon opening in vscode.
Currently, there are two effects taking care of the syncing process
between the VS Code file and KaotoUI (and source code) in the
KogitoEditorIntegrationProvider.tsx file. Said effects reacts when
one of the several dependencies changes, like flows or settings.

This causes a recursive call since the same effects are updating the
setting object from which they depends.

The fix is for the effect to rely only in the intended property.

This is still colliding with the Sourcecode sync that it's happening
in the Visualization.tsx component. A follow up will be needed to
consolidate both scenarios.

fixes: https://github.com/KaotoIO/vscode-kaoto/issues/285