fix(vscode): Missing backend version in vscode
Currently, there's a race condition caused between 'KogitoEditorIntegrationProvider.tsx'
and the 'KaotoToolbar.tsx' as the latter holds an outdated reference
to the settings object, causing to the backend version be removed.

The fix is to use the correct API for the 'setSettings()' hook in which
we only pass the property to be updated and nothing more.

fixes: https://github.com/KaotoIO/vscode-kaoto/issues/287