fix(vscode): Provide missing capabilities for @kie-core-tools
Currently, we're fetching the capabilities once upon loading the app and
later on we iterate over the already fetched array to save HTTP calls.

This presents a problem for our integration with @kie-core-tools as the
capabilities are fetched in a child component (AppLayout.tsx) of the
provider component (KaotoEditor.tsx).

The solution is to fetch the capabilities in an earlier phase, in the
new CapabilitiesProvider (capabilities.provider.tsx) to ensure the information
being available.

fixes: https://github.com/KaotoIO/vscode-kaoto/issues/293