fix(viz): Use metadata.name as flow identifier
Currently, the UI creates the flows IDs in two places:
1. Upon syncing the code through the useFlowsStore.setFlowsWrapper() method
2. Upon creating a new flow when using the New Flow button

For the .1, since the UI doesn't have a previous ID
because we're starting from the source code, the UI
creates an ID using the DSL name and the position index,
i.e. "Camel Route-1".

For the .2, the UI uses a service to generate an ID
with the word "route-####" while "####" it's a 4 digits
random number.

In this commit, the UI leverages the IFlowsWrapper.flows[0].metadata.name
property as an ID for each flow, this way, whenever a sync happens,
the IDs will be the same.

fixes: https://github.com/KaotoIO/vscode-kaoto/issues/280
fixes: https://github.com/KaotoIO/kaoto-ui/issues/1910