fix(flows): New route button should use supportsMultipleFlows property
Currently, there's a `supportsMultipleFlows` property from the capabilities
endpoint from the backend that the UI is not leveraging yet.

This property is meant to signal the UI whether a DSL supports multiple
flows or not.

This pull request makes use of the said property to disable the
New route button in case the UI is dealing with those kinds of DSL.

At the moment of writing this code, those DSLs are `Kamelet` and `KameletBindings`

fixes: https://github.com/KaotoIO/kaoto-ui/issues/2055