chore(e2e): Fix broken e2e
With the introduction of https://github.com/KaotoIO/kaoto-backend/pull/734,
the multiple flows support for Integration CRDs is no longer an individual
CRD per flow but rather a single CRD with many flows in it.

This commit takes that into consideration by updating the fixtures and the
associated lines.

In addition to that, the editorDeleteLine was refactored to leverage the
monaco GoToLine functionality to consistently reach a given line.