fix(steps): Cannot update steps that doesn't contain parameters
Currently, if a step doesn't contain parameters, like the do/try
step, is ignored in the updateStepParameters method.

The fix is to fill in the parameters with the newValues object if
provided, and from that point forward, insert the step in the store
since a step extension could modify it.

fixes: https://github.com/KaotoIO/kaoto-ui/issues/2051