fix(router): Support dynamic urls for kaoto-standalone
currently, Kaoto's UI is using the `KAOTO_API` environment
variable to determine the base URL for the API. This environment
variable gets loaded by using `dotenv-webpack` at building time.

In addition to that, the `index.html` file used `<base href="/">`
to signal where the UI should be loaded from, the problem with
this is that when deploying Kaoto in k8s or OCP platforms, the
URLs are gonna be dynamic and not `/`.

This commit makes a few adjustments, starting with removing
`dotenv-webpack` and replacing it with `dotenv` which is later
on leveraged in the `webpack.dev.js` file. The reason for this
is to leverage the local `.env` file only in `DEV` mode.

For `production` builds, the `KAOTO_API` variable needs to be set
by hand to allow both `dev` and `prod` builds locally.

The other relevant change was to move from `BrowserRouter` to `HashRouter`,
in order to provide a simple alternative to load Kaoto's UI from
dynamic routes since the `basePath` is unknown at the build time.

relates to: https://github.com/KaotoIO/kaoto-standalone/issues/32