chore: Convert project into monorepo
In the context of https://github.com/KaotoIO/kaoto-ui/discussions/1759,
and with the aim to improve the vscode-kaoto testing experience,
this commit turns the project into a monorepo.

This will bring a few challenges around prefixing yarn commands with
the right package, but at the same time, it will allow us to bring
other packages in an easy way.