feat(code-editor): Update icon when editor is dirty
Currently, when the source code is changed through the source code
editor, the sync button is always green, being the only feedback that
it was changed the top banner.

This commit updates the icon from the green check to an orange sync icon
whenever the code editor is dirty. It also changes the tooltip text from
"Sync your code" to "Code is synced"

fixes: https://github.com/KaotoIO/kaoto-ui/issues/2185