fix(sandbox): Build sandbox application with webpack
At this moment, vite's module federation plugin seems incompatible
with webpack's counterpart, causing the steps extensions not to load.

This commit adds a webpack config to build the sandbox application
without removing vite. The reasoning is that vite offers a much faster
experience and it's likely that the steps extensions won't be loaded
as module federation but rather by being part of the final build.

fixes: https://github.com/KaotoIO/kaoto-ui/issues/2216
relates: https://github.com/KaotoIO/kaoto-ui/discussions/1759